added agency field to users

+++ src/main/java/com/pfe/ecredit/service/TypeDocumentsService.java
@@ -7,7 +7,7 @@ import com.pfe.ecredit.domain.SiTypeDocuments;
 public interface TypeDocumentsService {
 	
 	public List<SiTypeDocuments> findAllDocuments();
-	public SiTypeDocuments findDocument(Long id);
-	public List<SiTypeDocuments> findDocumentsByCredit(Long id);
+	public SiTypeDocuments findDocument(Integer id);
+	public List<SiTypeDocuments> findDocumentsByCredit(Integer id);
 
 }
